Extended configuration to fix packet sizes. In future going to make configuration extendable, and network specific.
Added helper functions to convert some types to string. Helpful for debugging.

Added SlaveDescriptor class. This is a class to work with a specific slave. At the moment it IS NOT thread safe.

Anned class Master. This class manages all conencted slaves. And responsible for authentification, managing connection descriptors.

Notifying a slave about upcoming droping connection. The slave is expected to re-authentificate if it wants to work with the master again.

Fixed important bugs in UnixTCPSocketIO. Mostly setting necessary options to the connection socket.

Some minor class interface/documentation improvements.